+++ docs/source/history.md
@@ -376,3 +376,10 @@ Enjoy - we are awaiting comments and issues, thanks for testing.
   - New notebook: `sequential.ipynb`, creating an animation showing state-by-state wake propagation
 
 **Full Changelog**: [https://github.com/FraunhoferIWES/foxes/commits/v0.5.1](https://github.com/FraunhoferIWES/foxes/commits/v0.5.1)
+
+## v0.5.2
+
+- Vertical profiles:
+  - Now optional support for `ustar` input data instead of reference data
+
+**Full Changelog**: [https://github.com/FraunhoferIWES/foxes/commits/v0.5.2](https://github.com/FraunhoferIWES/foxes/commits/v0.5.2)